Fix database

+++ database/migrations/2021_06_18_004840_create_menu_table.php
@@ -19,7 +19,7 @@ class CreateMenuTable extends Migration
             $table->string('category', 50);
             $table->string('picture');
             $table->string('description');
-            $table->decimal('price', 10, 0)->default(0);
+            $table->integer('price')->default(0);
             $table->timestamp('created_at')->useCurrent();
             $table->timestamp('updated_at')->useCurrent();
         });
